Create modelAudit object

Function audit create modelAudit object for further validation of a model. Models may have very different structures. This function creates a unified representation of a model and calculates residuals, which can be further processed by various error analysis functions.

Function 'audit()' is deprecated, please, use an object of class 'explainer' created with function explain from the DALEX package.

audit(
  object,
  data = NULL,
  y = NULL,
  predict.function = NULL,
  residual.function = NULL,
  label = NULL
)

Arguments

object
An object containing a model or object of class explainer (see explain).

data
Data.frame or matrix - data that will be used by further validation functions. If not provided, will be extracted from the model.

y
Response vector that will be used by further validation functions. Some functions may require an integer vector containing binary labels with values 0,1. If not provided, will be extracted from the model.

predict.function
Function that takes two arguments: model and data. It should return a numeric vector with predictions.

residual.function
Function that takes three arguments: model, data and response vector. It should return a numeric vector with model residuals for given data. If not provided, response residuals (\(y-\hat{y}\)) are calculated.

label
Character - the name of the model. By default it's extracted from the 'class' attribute of the model.

Value

An object of class ModelAudit, which contains: #'

model.class class of the audited model,

label the name of the model,

model the audited model,

fitted.values fitted values from model,

data data used for fitting the model,

y vector with values of predicted variable used for fitting the model,

predict.function function that were used for model predictions,

residual.function function that were used for calculating model residuals,

residuals

std.residuals standardized residuals - the residuals divided by theirs standard deviation.

Examples

titanic <- na.omit(DALEX::titanic)
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  2099  rows  9  cols  extracted from the model  
#>   -> target variable   :  not specified! (  WARNING  )
#>   -> model_info        :  package stats , ver. 4.0.0 , task classification (  default  ) 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  9.814966e-09 , mean =  0.3244402 , max =  1  
#>   -> residual function :  difference between y and yhat (  default  )
#>   A new explainer has been created!  

p_fun <- function(model, data){predict(model, data, response = "link")}
audit_glm_newpred <- audit(model_glm, predict.function = p_fun)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  2099  rows  9  cols  extracted from the model  
#>   -> target variable   :  not specified! (  WARNING  )
#>   -> model_info        :  package stats , ver. 4.0.0 , task classification (  default  ) 
#>   -> predict function  :  predict.function 
#>   -> predicted values  :  numerical, min =  -18.43936 , mean =  -1.259491 , max =  17.34929  
#>   -> residual function :  difference between y and yhat (  default  )
#>   A new explainer has been created!  

library(randomForest)

#> randomForest 4.6-14

#> Type rfNews() to see new features/changes/bug fixes.

model_rf <- randomForest(Species ~ ., data=iris)
audit_rf <- audit(model_rf)

#> Preparation of a new explainer is initiated
#>   -> model label       :  randomForest  (  default  )
#>   -> data              :  150  rows  5  cols  extracted from the model  
#>   -> target variable   :  not specified! (  WARNING  )
#>   -> model_info        :  package randomForest , ver. 4.6.14 , task classification (  default  ) 
#>   -> predict function  :  yhat.randomForest  will be used (  default  )
#>   -> predicted values  :  predict function returns multiple columns:  3  (  WARNING  ) some of functionalities may not work 
#>   -> residual function :  difference between y and yhat (  default  )
#>   A new explainer has been created!