Automated tests for model residuals

Currently three tests are performed - for outliers in residuals - for autocorrelation in target variable or in residuals - for trend in residuals as a function of target variable (detection of bias)

check_residuals(object, ...)

Arguments

object
An object of class 'explainer' created with function explain from the DALEX package.

...
other parameters that will be passed to further functions.

Value

list with statistics for particualar checks

Examples

library(DALEX)

#> Welcome to DALEX (version: 0.4.6).
#> Find examples and detailed introduction at: https://pbiecek.github.io/PM_VEE/

#> 
#> Dołączanie pakietu: 'DALEX'

#> Następujące obiekty zostały zakryte z 'package:auditor':
#> 
#>     model_performance, theme_drwhy

dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_exp <- explain(lm_model, data = dragons, y = dragons$life_length)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (default)
#>   -> data              :  100  rows  8  cols 
#>   -> target variable   :  100  values 
#>   -> predict function  :  yhat.lm  will be used (default)
#>   -> predicted values  :  numerical, min =  585.8311 , mean =  1347.787 , max =  2942.307  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -88.41755 , mean =  -1.489291e-13 , max =  77.92805  
#> A new explainer has been created!

library(auditor)
check_residuals(lm_exp)

#>   -----------------------------------------------
#>    Checks for autocorrelation
#>   -----------------------------------------------
#>     Model name:  lm 
#>     Autocorrelation in target:     -0.08      
#>     Autocorrelation in residuals:  +0.06      
#>   -----------------------------------------------
#>    Checks for outliers
#>   -----------------------------------------------
#>     Model name:  lm 
#>     Shift > 1:  0 ( 0 %) 
#>     Shift > 2:  0 ( 0 %) 
#>     Top lowest standardised residuals: 
#>      -2.2738 (48), -2.2542 (90), -2.0906 (63), -2.0201 (45), -1.8596 (67) 
#>     Top highest standardised residuals: 
#>      2.004 (96), 1.9057 (2), 1.6531 (11), 1.4571 (3), 1.4286 (42) 
#>   -----------------------------------------------
#>    Checks for trend in residuals
#>   -----------------------------------------------
#>     Model name:  lm 
#>     Standardised loess fit:  +0.54      

 

 library("ranger")

#> 
#> Dołączanie pakietu: 'ranger'

#> Następujący obiekt został zakryty z 'package:randomForest':
#> 
#>     importance

 rf_model <- ranger(life_length ~ ., data = dragons)
 predict_function <- function(m,x,...) predict(m, x, ...)$predictions
 rf_exp <- explain(rf_model, data = dragons, y = dragons$life_length,
           predict_function = predict_function)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (default)
#>   -> data              :  100  rows  8  cols 
#>   -> target variable   :  100  values 
#>   -> predict function  :  predict_function 
#>   -> predicted values  :  numerical, min =  756.5293 , mean =  1341.695 , max =  2471.066  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -182.0128 , mean =  6.092016 , max =  427.355  
#> A new explainer has been created!

 check_residuals(rf_exp)

#>   -----------------------------------------------
#>    Checks for autocorrelation
#>   -----------------------------------------------
#>     Model name:  ranger 
#>     Autocorrelation in target:     -0.08      
#>     Autocorrelation in residuals:  -0.03      
#>   -----------------------------------------------
#>    Checks for outliers
#>   -----------------------------------------------
#>     Model name:  ranger 
#>     Shift > 1:  2 ( 2 %) 
#>     Shift > 2:  0 ( 0 %) 
#>     Top lowest standardised residuals: 
#>      -1.8651 (93), -1.8295 (95), -1.7291 (85), -1.4367 (12), -1.2446 (67) 
#>     Top highest standardised residuals: 
#>      4.177 (55), 3.7552 (26), 3.0359 (99), 2.3009 (75), 2.1753 (66) 
#>   -----------------------------------------------
#>    Checks for trend in residuals
#>   -----------------------------------------------
#>     Model name:  ranger 
#>     Standardised loess fit:  +13.71     *