Cook's distances

Calculates Cook's distances for each observation. Please, note that it will work only for functions with specified 'update' method.

model_cooksdistance(object)

observationInfluence(object)

Arguments

object
An object of class 'explain' created with function explain from the DALEX package.

Value

An object of class 'auditor_model_cooksdistance'.

References

Cook, R. Dennis (1977). "Detection of Influential Observations in Linear Regression". doi:10.2307/1268249.

Examples

titanic <- na.omit(DALEX::titanic)

# fit a model
model_glm <- glm(survived ~ ., family = binomial, data = titanic)

# use DALEX package to wrap up a model into explainer
exp_glm <- DALEX::explain(model_glm, data = titanic, y = titanic$survived)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  2099  rows  9  cols 
#>   -> target variable   :  2099  values 
#>   -> target variable   :  Please note that 'y' is a factor.  (  WARNING  )
#>   -> target variable   :  Consider changing the 'y' to a logical or numerical vector.
#>   -> target variable   :  Otherwise I will not be able to calculate residuals or loss function.
#>   -> data              :  A column identical to the target variable `y` has been found in the `data`.  (  WARNING  )
#>   -> data              :  It is highly recommended to pass `data` without the target variable column
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  9.814966e-09 , mean =  0.3244402 , max =  1  
#>   -> residual function :  difference between y and yhat (  default  )

#> Warning: ‘-’ not meaningful for factors

#>   -> residuals         :  numerical, min =  NA , mean =  NA , max =  NA  
#>   A new explainer has been created!  

# validate a model with auditor
library(auditor)
model_cooksdistance(exp_glm)

#> Model label:  lm 
#> 
#> Observations with the largest Cook's distances:
#>     Cook's distance index
#> 924    7.677289e+19   848
#> 870    3.214367e+19   795
#> 522    9.071796e+18   481
#> 
#> Quantiles of Cook's distances:
#>           0%          10%          20%          30%          40%          50% 
#> 1.735978e-12 6.090284e-06 1.005814e-05 1.424874e-05 2.053477e-05 3.426320e-05 
#>          60%          70%          80%          90%         100% 
#> 7.991737e-05 1.744884e-04 2.837646e-04 7.210318e-04 7.677289e+19