Create Halfnormal Explanation

Creates 'auditor_model_halfnormal' object that can be used for plotting halfnormal plot.

model_halfnormal(object, quant = FALSE, ...)

modelFit(object, quant = FALSE, ...)

Arguments

object
An object of class 'explainer' created with function explain from the DALEX package.

quant
if TRUE values on axis are on quantile scale.

...
other parameters passed do hnp function.

Value

An object of the class 'auditor_model_halfnormal'.

References

Moral, R., Hinde, J., & Demétrio, C. (2017). Half-Normal Plots and Overdispersed Models in R: The hnp Package.doi:http://dx.doi.org/10.18637/jss.v081.i10

Examples

titanic <- na.omit(DALEX::titanic[1:100,])

# fit a model
model_glm <- glm(survived ~ ., family = binomial, data = titanic)

# use DALEX package to wrap up a model into explainer
exp_glm <- DALEX::explain(model_glm)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  98  rows  9  cols (extracted from the model)
#>   -> target variable   :  not specified! (  WARNING  )
#>   -> predict function  :  yhat.glm  will be used (  default  )

#> Warning: prediction from a rank-deficient fit may be misleading

#>   -> predicted values  :  numerical, min =  1.721693e-09 , mean =  0.3673469 , max =  1  
#>   -> residual function :  difference between y and yhat (  default  )
#>   A new explainer has been created!  

# validate a model with auditor
library(auditor)
model_halfnormal(exp_glm)

#> Binomial model 

#> Model label:  
#> Quantiles of Residuals:
#>            0%           10%           20%           30%           40% 
#> -2.0268310187 -0.9891191280 -0.7507466977 -0.6184924338 -0.4052881806 
#>           50%           60%           70%           80%           90% 
#> -0.1205808471 -0.0001199433  0.5731707308  0.7043099673  1.1881626328 
#>          100% 
#>  2.2775116447