Influence of observations Plot in D3 with r2d3 Package.

Cook’s distances are used for estimate the influence of an single observation. If the picture is not displayed in the viewer, please update your RStudio.

plotD3CooksDistance(object, ..., nlabel = 3, single_plot = FALSE,
  scale_plot = FALSE, background = FALSE)

Arguments

object
An object of class modelAudit or observationInfluence.

...
Other modelAudit or modelResiduals objects to be plotted together.

nlabel
Number of observations with the biggest Cook's distances to be labeled.

single_plot
Logical, indicates whenever single or facets should be plotted. By default it's FALSE.

scale_plot
Logical, indicates whenever the plot should scale with height. By default it's FALSE.

background
Logical, available only if single_plot = FALSE. Indicates whenever backgroud plots should be plotted. By default it's FALSE.

See also

Examples

library("auditor")
library(car)

lm_model <- lm(prestige~education + women + income, data = Prestige)
lm_au <- audit(lm_model, data = Prestige, y = Prestige$prestige, label = "lm")
plotD3CooksDistance(lm_au)

library(randomForest)
rf_model <- randomForest(prestige~education + women + income, data = Prestige)
rf_au <- audit(rf_model, data = Prestige, y = Prestige$prestige, label = "rf")
plotD3CooksDistance(lm_au, rf_au, nlabel = 4)

#> 1 out of 102 2 out of 102 3 out of 102 4 out of 102 5 out of 102 6 out of 102 7 out of 102 8 out of 102 9 out of 102 10 out of 102 11 out of 102 12 out of 102 13 out of 102 14 out of 102 15 out of 102 16 out of 102 17 out of 102 18 out of 102 19 out of 102 20 out of 102 21 out of 102 22 out of 102 23 out of 102 24 out of 102 25 out of 102 26 out of 102 27 out of 102 28 out of 102 29 out of 102 30 out of 102 31 out of 102 32 out of 102 33 out of 102 34 out of 102 35 out of 102 36 out of 102 37 out of 102 38 out of 102 39 out of 102 40 out of 102 41 out of 102 42 out of 102 43 out of 102 44 out of 102 45 out of 102 46 out of 102 47 out of 102 48 out of 102 49 out of 102 50 out of 102 51 out of 102 52 out of 102 53 out of 102 54 out of 102 55 out of 102 56 out of 102 57 out of 102 58 out of 102 59 out of 102 60 out of 102 61 out of 102 62 out of 102 63 out of 102 64 out of 102 65 out of 102 66 out of 102 67 out of 102 68 out of 102 69 out of 102 70 out of 102 71 out of 102 72 out of 102 73 out of 102 74 out of 102 75 out of 102 76 out of 102 77 out of 102 78 out of 102 79 out of 102 80 out of 102 81 out of 102 82 out of 102 83 out of 102 84 out of 102 85 out of 102 86 out of 102 87 out of 102 88 out of 102 89 out of 102 90 out of 102 91 out of 102 92 out of 102 93 out of 102 94 out of 102 95 out of 102 96 out of 102 97 out of 102 98 out of 102 99 out of 102 100 out of 102 101 out of 102 102 out of 102