Plot Prediction vs Target, Observed or Variable Values in D3 with r2d3 package.

Function plotD3_prediction plots predicted values observed or variable values in the model.

plotD3_prediction(
  object,
  ...,
  variable = "_y_",
  points = TRUE,
  smooth = FALSE,
  abline = FALSE,
  point_count = NULL,
  single_plot = TRUE,
  scale_plot = FALSE,
  background = FALSE
)

plotD3Prediction(
  object,
  ...,
  variable = NULL,
  points = TRUE,
  smooth = FALSE,
  abline = FALSE,
  point_count = NULL,
  single_plot = TRUE,
  scale_plot = FALSE,
  background = FALSE
)

Arguments

object
An object of class 'auditor_model_residual.

...
Other modelAudit or modelResiduals objects to be plotted together.

variable
Name of variable to order residuals on a plot. If variable="_y_", the data is ordered by a vector of actual response (y parameter passed to the explain function). If variable = "_y_hat_" the data on the plot will be ordered by predicted response. If variable = NULL, unordered observations are presented.

points
Logical, indicates whenever observations should be added as points. By default it's TRUE.

smooth
Logical, indicates whenever smoothed lines should be added. By default it's FALSE.

abline
Logical, indicates whenever function y = x should be added. Works only with variable = NULL which is a default option.

point_count
Number of points to be plotted per model. Points will be chosen randomly. By default plot all of them.

single_plot
Logical, indicates whenever single or facets should be plotted. By default it's TRUE.

scale_plot
Logical, indicates whenever the plot should scale with height. By default it's FALSE.

background
Logical, available only if single_plot = FALSE. Indicates whenever background plots should be plotted. By default it's FALSE.

Value

a r2d3 object

See also

Examples

dragons <- DALEX::dragons[1:100, ]

# fit a model
model_lm <- lm(life_length ~ ., data = dragons)

lm_audit <- audit(model_lm, data = dragons, y = dragons$life_length)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  100  rows  8  cols 
#>   -> target variable   :  100  values 
#>   -> predict function  :  yhat.lm  will be used (  default  )
#>   -> predicted values  :  No value for predict function target column. (  default  )
#>   -> model_info        :  package stats , ver. 4.1.1 , task regression (  default  ) 
#>   -> predicted values  :  numerical, min =  585.8311 , mean =  1347.787 , max =  2942.307  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -88.41755 , mean =  -1.489291e-13 , max =  77.92805  
#>   A new explainer has been created!  

# validate a model with auditor
mr_lm <- model_residual(lm_audit)

# plot results
plotD3_prediction(mr_lm, abline = TRUE)
plotD3_prediction(mr_lm, variable = "height", smooth = TRUE)

library(randomForest)
model_rf <- randomForest(life_length~., data = dragons)
rf_audit <- audit(model_rf, data = dragons, y = dragons$life_length)

#> Preparation of a new explainer is initiated
#>   -> model label       :  randomForest  (  default  )
#>   -> data              :  100  rows  8  cols 
#>   -> target variable   :  100  values 
#>   -> predict function  :  yhat.randomForest  will be used (  default  )
#>   -> predicted values  :  No value for predict function target column. (  default  )
#>   -> model_info        :  package randomForest , ver. 4.6.14 , task regression (  default  ) 
#>   -> predicted values  :  numerical, min =  775.5201 , mean =  1343.438 , max =  2499.293  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -187.285 , mean =  4.348868 , max =  399.1275  
#>   A new explainer has been created!  

mr_rf <- model_residual(rf_audit)
plotD3_prediction(mr_lm, mr_rf, variable = "weight", smooth = TRUE)