Half-Normal plot

The half-normal plot is one of the tools designed to evaluate the goodness of fit of a statistical models. It is a graphical method for comparing two probability distributions by plotting their quantiles against each other. Points on the plot correspond to ordered absolute values of model diagnostic (i.e. standardized residuals) plotted against theoretical order statistics from a half-normal distribution.

plotHalfNormal(object, score = TRUE, quant.scale = FALSE,
  xlab = "half-normal Quantiles", ylab = "residuals", main = "", ...)

Arguments

object
ModelAudit object, fitted model object or numeric vector.

score
If TRUE score based on probability density function is displayed on the plot.

quant.scale
if TRUE values on axis are on quantile scale.

xlab
The text for the x axis.

ylab
The text for the y axis.

main
Title of plot.

...
extra arguments passed to hnp.

Value

An object of class ggplot

See also