Area Under ROC Curve (AUC)

Area Under Curve (AUC) for Receiver Operating Characteristic.

score_auc(object, data = NULL)

scoreROC(object)

Arguments

object
An object of class explainer created with function explain from the DALEX package.

data
New data that will be used to calcuate the score. Pass NULL if you want to use data from object.

Value

An object of class auditor_score.

See also

Examples

titanic <- na.omit(DALEX::titanic)
titanic$survived <- titanic$survived == "yes"

# fit a model
model_glm <- glm(survived ~ ., family = binomial, data = titanic)

#create an explainer
exp_glm <- DALEX::explain(model_glm, y = titanic$survived)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  2099  rows  9  cols (extracted from the model)
#>   -> target variable   :  2099  values 
#>   -> data              :  A column identical to the target variable `y` has been found in the `data`.  (  WARNING  )
#>   -> data              :  It is highly recommended to pass `data` without the target variable column
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  9.814966e-09 , mean =  0.3244402 , max =  1  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.9614217 , mean =  -1.68201e-09 , max =  0.9666502  
#>   A new explainer has been created!  

# calculate score
score_auc(exp_glm)

#> auc: 0.8293568