Half-Normal Score

Score is approximately: \( \sum{\#[res_i \leq simres_{i,j}] - n } \) with the distinction that each element of sum is also scaled to take values from [0,1].

\(res_i\) is a residual for i-th observation, \(simres_{i,j}\) is the residual of j-th simulation for i-th observation, and \(n\) is the number of simulations for each observation. Scores are calculated on the basis of simulated data, so they may differ between function calls.

score_halfnormal(object, ...)

scoreHalfNormal(object, ...)

Arguments

object
An object of class explainer created with function explain from the DALEX package.

...
...

Value

An object of class auditor_score.

Examples

dragons <- DALEX::dragons[1:100, ]

# fit a model
model_lm <- lm(life_length ~ ., data = dragons)

# create an explainer
exp_lm <- DALEX::explain(model_lm, data = dragons, y = dragons$life_length)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  100  rows  8  cols 
#>   -> target variable   :  100  values 
#>   -> model_info        :  package stats , ver. 4.0.0 , task regression (  default  ) 
#>   -> predict function  :  yhat.lm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  585.8311 , mean =  1347.787 , max =  2942.307  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -88.41755 , mean =  -1.489291e-13 , max =  77.92805  
#>   A new explainer has been created!  

# calculate score
score_halfnormal(exp_lm)

#> Gaussian model (lm object) 

#> halfnormal: NA