Influence of observations Plot in D3 with r2d3 package.

Cook’s distances are used for estimate the influence of an single observation. If the picture is not displayed in the viewer, please update your RStudio.

plotD3CooksDistance(object, ..., nlabel = 3, single_plot = FALSE,
  scale_plot = FALSE, background = FALSE)

Arguments

object
An object of class modelAudit or observationInfluence.

...
Other modelAudit or modelResiduals objects to be plotted together.

nlabel
Number of observations with the biggest Cook's distances to be labeled.

single_plot
Logical, indicates whenever single or facets should be plotted. By default it's FALSE.

scale_plot
Logical, indicates whenever the plot should scale with height. By default it's FALSE.

background
Logical, available only if single_plot = FALSE. Indicates whenever backgroud plots should be plotted. By default it's FALSE.

Value

a `r2d3` object.

See also

Examples

library("auditor")
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
plotD3CooksDistance(lm_au)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length, label = "rf")
plotD3CooksDistance(lm_au, rf_au, nlabel = 4)

#> 1 out of 100 2 out of 100 3 out of 100 4 out of 100 5 out of 100 6 out of 100 7 out of 100 8 out of 100 9 out of 100 10 out of 100 11 out of 100 12 out of 100 13 out of 100 14 out of 100 15 out of 100 16 out of 100 17 out of 100 18 out of 100 19 out of 100 20 out of 100 21 out of 100 22 out of 100 23 out of 100 24 out of 100 25 out of 100 26 out of 100 27 out of 100 28 out of 100 29 out of 100 30 out of 100 31 out of 100 32 out of 100 33 out of 100 34 out of 100 35 out of 100 36 out of 100 37 out of 100 38 out of 100 39 out of 100 40 out of 100 41 out of 100 42 out of 100 43 out of 100 44 out of 100 45 out of 100 46 out of 100 47 out of 100 48 out of 100 49 out of 100 50 out of 100 51 out of 100 52 out of 100 53 out of 100 54 out of 100 55 out of 100 56 out of 100 57 out of 100 58 out of 100 59 out of 100 60 out of 100 61 out of 100 62 out of 100 63 out of 100 64 out of 100 65 out of 100 66 out of 100 67 out of 100 68 out of 100 69 out of 100 70 out of 100 71 out of 100 72 out of 100 73 out of 100 74 out of 100 75 out of 100 76 out of 100 77 out of 100 78 out of 100 79 out of 100 80 out of 100 81 out of 100 82 out of 100 83 out of 100 84 out of 100 85 out of 100 86 out of 100 87 out of 100 88 out of 100 89 out of 100 90 out of 100 91 out of 100 92 out of 100 93 out of 100 94 out of 100 95 out of 100 96 out of 100 97 out of 100 98 out of 100 99 out of 100 100 out of 100