Model Scores computations

This function provides several scores for model validation and performance assessment. Scores can be also used to compare models.

score(object, score = "mse", data = NULL, ...)

Arguments

object
An object of class explainer created with function explain from the DALEX package.

score
The score to be calculated. Possible values: acc, auc, cookdistance, dw, f1, halfnormal, mae, mse, peak, precision, rec, recall, rmse, rroc, runs, specificity, one_minus_acc, one_minus_auc, one_minus_f1, one_minus_precision, one_minus_recall, one_minus_specificity (for detailed description see functions in see also section).

data
New data that will be used to calcuate the score. Pass NULL if you want to use data from object.

...
Other arguments dependent on the type of score.

Value

An object of class auditor_score, except Cooks distance, where numeric vector is returned.

See also

Examples

dragons <- DALEX::dragons[1:100, ]

# fit a model
model_lm <- lm(life_length ~ ., data = dragons)

# use DALEX package to wrap up a model into explainer
exp_lm <- DALEX::explain(model_lm, data = dragons, y = dragons$life_length)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  100  rows  8  cols 
#>   -> target variable   :  100  values 
#>   -> predict function  :  yhat.lm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  585.8311 , mean =  1347.787 , max =  2942.307  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -88.41755 , mean =  -1.489291e-13 , max =  77.92805  
#>   A new explainer has been created!  

# calculate score
score(exp_lm, score = 'mae')

#> mae: 31.81926